Allow husl to be unbundled.
