+++ vispy/ext/husl.py
@@ -1,331 +1,12 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2015, Vispy Development Team.
+# Distributed under the (new) BSD License. See LICENSE.txt for more info.
+
 """
-HUSL colors python implementation.
-
-Source: https://github.com/husl-colors/husl.py
-
-Copyright (c) 2015 Alexei Boronine
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
+Handle loading husl package from system or from the bundled copy
 """
 
-import operator
-import math
-
-__version__ = "4.0.2"
-
-
-m = [
-    [3.240969941904521, -1.537383177570093, -0.498610760293],
-    [-0.96924363628087, 1.87596750150772, 0.041555057407175],
-    [0.055630079696993, -0.20397695888897, 1.056971514242878],
-]
-m_inv = [
-    [0.41239079926595, 0.35758433938387, 0.18048078840183],
-    [0.21263900587151, 0.71516867876775, 0.072192315360733],
-    [0.019330818715591, 0.11919477979462, 0.95053215224966],
-]
-refX = 0.95045592705167
-refY = 1.0
-refZ = 1.089057750759878
-refU = 0.19783000664283
-refV = 0.46831999493879
-kappa = 903.2962962
-epsilon = 0.0088564516
-
-
-# Public API
-
-def husl_to_rgb(h, s, l):
-    return lch_to_rgb(*husl_to_lch([h, s, l]))
-
-
-def husl_to_hex(h, s, l):
-    return rgb_to_hex(husl_to_rgb(h, s, l))
-
-
-def rgb_to_husl(r, g, b):
-    return lch_to_husl(rgb_to_lch(r, g, b))
-
-
-def hex_to_husl(hex):
-    return rgb_to_husl(*hex_to_rgb(hex))
-
-
-def huslp_to_rgb(h, s, l):
-    return lch_to_rgb(*huslp_to_lch([h, s, l]))
-
-
-def huslp_to_hex(h, s, l):
-    return rgb_to_hex(huslp_to_rgb(h, s, l))
-
-
-def rgb_to_huslp(r, g, b):
-    return lch_to_huslp(rgb_to_lch(r, g, b))
-
-
-def hex_to_huslp(hex):
-    return rgb_to_huslp(*hex_to_rgb(hex))
-
-
-def lch_to_rgb(l, c, h):
-    return xyz_to_rgb(luv_to_xyz(lch_to_luv([l, c, h])))
-
-
-def rgb_to_lch(r, g, b):
-    return luv_to_lch(xyz_to_luv(rgb_to_xyz([r, g, b])))
-
-
-def get_bounds(L):
-    sub1 = ((L + 16.0) ** 3.0) / 1560896.0
-    sub2 = sub1 if sub1 > epsilon else L / kappa
-    ret = []
-    for [m1, m2, m3] in m:
-        for t in [0, 1]:
-            top1 = (284517.0 * m1 - 94839.0 * m3) * sub2
-            top2 = ((838422.0 * m3 + 769860.0 * m2 + 731718.0 * m1)
-                    * L * sub2 - 769860.0 * t * L)
-            bottom = (632260.0 * m3 - 126452.0 * m2) * sub2 + 126452.0 * t
-            ret.append((top1 / bottom, top2 / bottom))
-    return ret
-
-
-def intersect_line_line(line1, line2):
-    return (line1[1] - line2[1]) / (line2[0] - line1[0])
-
-
-def distance_from_pole(point):
-    return math.sqrt(point[0] ** 2 + point[1] ** 2)
-
-
-def length_of_ray_until_intersect(theta, line):
-    m1, b1 = line
-    length = b1 / (math.sin(theta) - m1 * math.cos(theta))
-    if length < 0:
-        return None
-    return length
-
-
-def max_safe_chroma_for_L(L):
-    lengths = []
-    for [m1, b1] in get_bounds(L):
-        x = intersect_line_line((m1, b1), (-1.0 / m1, 0.0))
-        lengths.append(distance_from_pole((x, b1 + x * m1)))
-    return min(lengths)
-
-
-def max_chroma_for_LH(L, H):
-    hrad = H / 360.0 * math.pi * 2.0
-    lengths = []
-    for line in get_bounds(L):
-        l = length_of_ray_until_intersect(hrad, line)
-        if l is not None:
-            lengths.append(l)
-    return min(lengths)
-
-
-def dot_product(a, b):
-    return sum(map(operator.mul, a, b))
-
-
-def f(t):
-    if t > epsilon:
-        return 116 * math.pow((t / refY), 1.0 / 3.0) - 16.0
-    else:
-        return (t / refY) * kappa
-
-
-def f_inv(t):
-    if t > 8:
-        return refY * math.pow((t + 16.0) / 116.0, 3.0)
-    else:
-        return refY * t / kappa
-
-
-def from_linear(c):
-    if c <= 0.0031308:
-        return 12.92 * c
-    else:
-        return (1.055 * math.pow(c, 1.0 / 2.4) - 0.055)
-
-
-def to_linear(c):
-    a = 0.055
-
-    if c > 0.04045:
-        return (math.pow((c + a) / (1.0 + a), 2.4))
-    else:
-        return (c / 12.92)
-
-
-def rgb_prepare(triple):
-    ret = []
-    for ch in triple:
-        ch = round(ch, 3)
-
-        if ch < -0.0001 or ch > 1.0001:
-            raise Exception("Illegal RGB value %f" % ch)
-
-        if ch < 0:
-            ch = 0
-        if ch > 1:
-            ch = 1
-
-        # Fix for Python 3 which by default rounds 4.5 down to 4.0
-        # instead of Python 2 which is rounded to 5.0 which caused
-        # a couple off by one errors in the tests. Tests now all pass
-        # in Python 2 and Python 3
-        ret.append(round(ch * 255 + 0.001, 0))
-
-    return ret
-
-
-def hex_to_rgb(hex):
-    if hex.startswith('#'):
-        hex = hex[1:]
-    r = int(hex[0:2], 16) / 255.0
-    g = int(hex[2:4], 16) / 255.0
-    b = int(hex[4:6], 16) / 255.0
-    return [r, g, b]
-
-
-def rgb_to_hex(triple):
-    [r, g, b] = triple
-    return '#%02x%02x%02x' % tuple(rgb_prepare([r, g, b]))
-
-
-def xyz_to_rgb(triple):
-    xyz = map(lambda row: dot_product(row, triple), m)
-    return list(map(from_linear, xyz))
-
-
-def rgb_to_xyz(triple):
-    rgbl = list(map(to_linear, triple))
-    return list(map(lambda row: dot_product(row, rgbl), m_inv))
-
-
-def xyz_to_luv(triple):
-    X, Y, Z = triple
-
-    if X == Y == Z == 0.0:
-        return [0.0, 0.0, 0.0]
-
-    varU = (4.0 * X) / (X + (15.0 * Y) + (3.0 * Z))
-    varV = (9.0 * Y) / (X + (15.0 * Y) + (3.0 * Z))
-    L = f(Y)
-
-    # Black will create a divide-by-zero error
-    if L == 0.0:
-        return [0.0, 0.0, 0.0]
-
-    U = 13.0 * L * (varU - refU)
-    V = 13.0 * L * (varV - refV)
-
-    return [L, U, V]
-
-
-def luv_to_xyz(triple):
-    L, U, V = triple
-
-    if L == 0:
-        return [0.0, 0.0, 0.0]
-
-    varY = f_inv(L)
-    varU = U / (13.0 * L) + refU
-    varV = V / (13.0 * L) + refV
-    Y = varY * refY
-    X = 0.0 - (9.0 * Y * varU) / ((varU - 4.0) * varV - varU * varV)
-    Z = (9.0 * Y - (15.0 * varV * Y) - (varV * X)) / (3.0 * varV)
-
-    return [X, Y, Z]
-
-
-def luv_to_lch(triple):
-    L, U, V = triple
-
-    C = (math.pow(math.pow(U, 2) + math.pow(V, 2), (1.0 / 2.0)))
-    hrad = (math.atan2(V, U))
-    H = math.degrees(hrad)
-    if H < 0.0:
-        H = 360.0 + H
-
-    return [L, C, H]
-
-
-def lch_to_luv(triple):
-    L, C, H = triple
-
-    Hrad = math.radians(H)
-    U = (math.cos(Hrad) * C)
-    V = (math.sin(Hrad) * C)
-
-    return [L, U, V]
-
-
-def husl_to_lch(triple):
-    H, S, L = triple
-
-    if L > 99.9999999:
-        return [100, 0.0, H]
-    if L < 0.00000001:
-        return [0.0, 0.0, H]
-
-    mx = max_chroma_for_LH(L, H)
-    C = mx / 100.0 * S
-
-    return [L, C, H]
-
-
-def lch_to_husl(triple):
-    L, C, H = triple
-
-    if L > 99.9999999:
-        return [H, 0.0, 100.0]
-    if L < 0.00000001:
-        return [H, 0.0, 0.0]
-
-    mx = max_chroma_for_LH(L, H)
-    S = C / mx * 100.0
-
-    return [H, S, L]
-
-
-def huslp_to_lch(triple):
-    H, S, L = triple
-
-    if L > 99.9999999:
-        return [100, 0.0, H]
-    if L < 0.00000001:
-        return [0.0, 0.0, H]
-
-    mx = max_safe_chroma_for_L(L)
-    C = mx / 100.0 * S
-
-    return [L, C, H]
-
-
-def lch_to_huslp(triple):
-    L, C, H = triple
-
-    if L > 99.9999999:
-        return [H, 0.0, 100.0]
-    if L < 0.00000001:
-        return [H, 0.0, 0.0]
-
-    mx = max_safe_chroma_for_L(L)
-    S = C / mx * 100.0
-
-    return [H, S, L]
+try:
+    from ._bundled.husl import *  # noqa
+except ImportError:
+    from husl import *  # noqa
